243: re-org candid_rpc
