+++ src/candid_rpc/mod.rs
@@ -1,4 +1,6 @@
 mod cketh_conversion;
+#[cfg(test)]
+mod tests;
 
 use crate::rpc_client::{EthRpcClient, MultiCallError};
 use crate::{
@@ -163,77 +165,3 @@ fn get_transaction_hash(raw_signed_transaction_hex: &Hex) -> Option<Hex32> {
     Some(Hex32::from(transaction.hash.0))
 }
 
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::rpc_client::{MultiCallError, MultiCallResults};
-    use evm_rpc_types::{ProviderError, RpcError};
-
-    #[test]
-    fn test_process_result_mapping() {
-        use evm_rpc_types::{EthMainnetService, RpcService};
-
-        let method = RpcMethod::EthGetTransactionCount;
-
-        assert_eq!(
-            process_result(method, Ok(5)),
-            MultiRpcResult::Consistent(Ok(5))
-        );
-        assert_eq!(
-            process_result(
-                method,
-                Err(MultiCallError::<()>::ConsistentError(
-                    RpcError::ProviderError(ProviderError::MissingRequiredProvider)
-                ))
-            ),
-            MultiRpcResult::Consistent(Err(RpcError::ProviderError(
-                ProviderError::MissingRequiredProvider
-            )))
-        );
-        assert_eq!(
-            process_result(
-                method,
-                Err(MultiCallError::<()>::InconsistentResults(
-                    MultiCallResults::default()
-                ))
-            ),
-            MultiRpcResult::Inconsistent(vec![])
-        );
-        assert_eq!(
-            process_result(
-                method,
-                Err(MultiCallError::InconsistentResults(
-                    MultiCallResults::from_non_empty_iter(vec![(
-                        RpcService::EthMainnet(EthMainnetService::Ankr),
-                        Ok(5)
-                    )])
-                ))
-            ),
-            MultiRpcResult::Inconsistent(vec![(
-                RpcService::EthMainnet(EthMainnetService::Ankr),
-                Ok(5)
-            )])
-        );
-        assert_eq!(
-            process_result(
-                method,
-                Err(MultiCallError::InconsistentResults(
-                    MultiCallResults::from_non_empty_iter(vec![
-                        (RpcService::EthMainnet(EthMainnetService::Ankr), Ok(5)),
-                        (
-                            RpcService::EthMainnet(EthMainnetService::Cloudflare),
-                            Err(RpcError::ProviderError(ProviderError::NoPermission))
-                        )
-                    ])
-                ))
-            ),
-            MultiRpcResult::Inconsistent(vec![
-                (RpcService::EthMainnet(EthMainnetService::Ankr), Ok(5)),
-                (
-                    RpcService::EthMainnet(EthMainnetService::Cloudflare),
-                    Err(RpcError::ProviderError(ProviderError::NoPermission))
-                )
-            ])
-        );
-    }
-}
